+++ package.json
@@ -125,7 +125,6 @@
     },
     "dependencies": {
         "@babel/runtime": "^7.15.4",
-        "@metamask/eth-sig-util": "^4.0.0",
         "events": "^3.3.0",
         "cross-fetch": "^3.1.4"
     },
